Issue 418: Slowness when manipulating DOM because of test in callDomLoadFunctions
http://code.google.com/p/swfobject/issues/detail?id=418

There is an attempt to use document.getElementsByTagName("body")[0] before it is defined. Since try/catch is expensive, I have added a simple test if(!document.getElementsByTagName("body")[0]) to ensure <body> is available before attempting to add a <span>. Hopefully this clears up some of the slowness reported in this thread.

Signed-off-by: Philip Hutchison <philip@pipwerks.com>