Issue 533: Can only pass a selector to embedSWF
http://code.google.com/p/swfobject/issues/detail?id=533

Added ability to pass element reference in place of selector for embedSWF, createSWF and removeSWF functions. The main edit was in the getElementById 'helper' function, which means any function in the SWFObject library that uses getElementById will be able to accept a selector OR an element reference, including setVisibility and others.

Tested in:
Mac OS X: Chrome 11, Safari 5, Firefox 4, Opera 11
Win7: Firefox 3.6, IE9
WinXP: Firefox 3, IE7

Signed-off-by: Philip Hutchison <philip@pipwerks.com>