Changed IE handling in createSWF… again.
Turns out the new createSWF approach had a major bug: although adding
<param> nodes using createElement worked in IE, <param> elements were
ignored. For example, a <param> element containing flashvars would
never send flashvars to the SWF, even though the <param> node shows up
correctly in the DOM inspector. Unfortunately, IE requires param nodes
to be included during the object's creation, or they will be ignored
(no appendChild or innerHTML after the fact). Therefore, I reconfigured
the function to attach the <param> nodes to a <div> instead of an
<object>, continuing to take advantage of the benefits of createElement
(protection from script injection). The script then grabs the param
nodes from the div as a string (via innerHTML) and inserts the string
into the real <object> at the point of its creation. Funky, but it
works, and still enables us to avoid the older SWFObject string
concatenation approach in IE.