Semantics: replaced mentions of 'alt' content with 'fallback'
SWFObject has traditionally used the term "alternate content" instead
of "fallback content". To some, the two terms have identical meanings.
To others, they have different connotations that can be problematic.

SWFObject has always championed the notion of fallback content,
regardless of the term we assigned to the concept: markup will be
displayed if the SWF fails to load due to lack of plugin or lack of
supported plugin version. (Side note: SWFObject does not track whether
the SWF actually loads or not.)

I believe "fallback" is a more appropriate term moving forward, and as
indicated by the HTML5 docs, this is an appropriate choice:
"Some embedded content elements can have fallback content: content that
is to be used when the external resource cannot be used (e.g. because
it is of an unsupported format)."
http://www.w3.org/TR/html5/embedded-content-0.html

I have therefore edited the SWFObject library to replace all instances
of "alternate" with "fallback", and all shorthand "alt" with "fb".