Static publishing: Hiding the <object> sooner
When using static publishing and IE, if the visitor doesn't have the
proper version of Flash Player (and no ExpressInstall), the <object> is
removed and replaced by the fallback markup. In IE, the <object> is
hidden before it's removed due to load reference issues (as noted by
Bobby in the source code). Currently, the fallback content is extracted
from the <object> before the <object> is hidden. If the <object>
contains a large amount of fallback markup, this could conceivably
cause a 'flash' (forgive the pun) of content prior to the <object>'s
removal.  I have therefore moved the obj.style.display = "none" to
occur *before* the fallback content is extracted. This change has no
effect other than helping ensure the page doesn't suffer from a
FOUC-type event.