Simplified createCSS function
Removed check for IE on Macs since we no longer support it.
Simplified the 'add style rule' section to use feature detect and not
browser sniffing.
Removed the check for media in (media && typeof media == "string"); if
typeof media == "undefined" and not "string", the condition will
evaluate false the same as (media && typeof media == "string")